Implements logging
Uses winstone to create a logger which saves all requests to a file. Follows this tutorial: https://mirzaleka.medium.com/automated-logging-in-express-js-a1f85ca6c5cd.
Observe that the typescript types were removed in all added files and that clientIP breaks if it is left as an option in the httpRequestFormatter.